tests: Move Pipeline Binary to own test file
